+++ CHANGELOG.md
@@ -15,6 +15,7 @@ About changelog [here](https://keepachangelog.com/en/1.0.0/)
  - View multiple coverage tracks for samples within a case.
  - Renderings are debounced and keep track of the last render such that they not double-render
  - Tracks can be shown / hidden, collapsed and moved
+ - Tracks menu and customize Y-axis
 
 ### Changed
  - Resolution increased 2x for tracks.